pyproject.toml
[tool.poetry]
name = "concrexit"
version = "0.0.dev"
description = "New new Thalia website, now with extra Django."
authors = ["Thalia Technicie <www@thalia.nu>"]
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.9"
django-localflavor = "3.1"
freezegun = "1.2.2"
bleach = {extras = ["css"], version = "^5.0.0"}
djangorestframework = "3.14.0"
django-ical = "1.8.3"
django-libsass = "0.8.0"
python-magic = "0.4.27"
Pillow = "9.2.0"
Django = "4.1.2"
django-compressor = "4.1"
psycopg2 = {version = "2.9.4", optional = true}
bcrypt = "4.0.1"
argon2_cffi = "21.3.0"
firebase-admin = "6.0.1"
sentry-sdk = "1.9.10"
django-sendfile2 = "0.7.0"
django-queryable-properties = "1.8.3"
google-api-python-client = "2.65.0"
django-oauth-toolkit = "2.2.0"
django-cors-headers = "3.13.0"
django-debug-toolbar = "3.7.0"
django-admin-autocomplete-filter = "0.7.1"
PyYAML = "6.0"
pydocstyle = "6.1.1"
django-bootstrap5 = "^22.1"
django-tinymce = "^3.4.0"
beautifulsoup4 = "^4.11.1"
django-storages = "^1.12.3"
boto3 = "^1.23.7"
gunicorn = "^20.1.0"

# docs requirements
recommonmark = { version = "0.7.1", optional = true }
sphinx = { version = "5.3.0", optional = true }
qrcode = {version = "^7.3.1", extras = ["pil"]}

[tool.poetry.extras]
docs = ["recommonmark", "sphinx"]
postgres = ["psycopg2"]

[tool.poetry.dev-dependencies]
django-template-check = "0.4.0"
factory_boy = "3.2.1"
pydenticon = "0.3.1"
Faker = "15.1.1"
coverage = {extras = ["toml"], version = "^6.0.0"}
isort = "5.10.1"
black = "22.10.0"
pylint = "2.13.9"
pylint-django = "2.5.3"

[tool.coverage.run]
branch = true
source = ["website"]
omit = ["website/manage.py", "website/thaliawebsite/wsgi.py", "website/thaliawebsite/settings.py", "website/*/migrations/*", "website/*/test*"]

[tool.coverage.report]
precision = 2
skip_covered = true
show_missing = true
fail_under = 65.0

[tool.black]
exclude = '(/(\.eggs|\.git|\.tox)/|migrations)'

[tool.isort]
profile = "black"
skip_glob = ["*/migrations/*"]
sections = [
    "FUTURE",
    "STDLIB",
    "DJANGO",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]

known_django = ["django"]
known_first_party = [
    "website",
    "activemembers",
    "announcements",
    "documents",
    "education",
    "events",
    "mailinglists",
    "members",
    "merchandise",
    "newsletters",
    "partners",
    "payments",
    "photos",
    "pizzas",
    "promotion",
    "pushnotifications",
    "registrations",
    "sales",
    "shortlinks",
    "singlepages",
    "thabloid",
    "thaliawebsite",
    "utils",
]
